requirements.txt
kafka-python
redis
faker
pandas
scikit-learn
xgboost
fastapi
uvicorn